Escape usernames when building spi-user-info url.
Fixes #59, but there may be other places where this still needs to get fixed.

Also, escapes "/" as "%252F" instead of the simpler "%2F", because
this is making one trip through quote_plus(), and another trip through
django's template autoescape mechanism.  This looks like it's more
trouble to fix than it's worth, as it's only a cosmetic issue.