Handle permissiondenied APIErrors better.
Instead of catching the error at the view level and displaying an
error message, we now catch it at the wiki_interface level and
silently return no data.  This allows non-admins to at least see the
anonymously-available data.

The disadvantage is that it's hard to do clean messaging.  What would
be nice is if we could put up a message saying, "Here's what you can
see; you don't have rights to see what's deleted).  To fix that, we'd
have to rejigger the entire pipeline to (for example) pass around
(status, data) pairs.  It doesn't seem worth the effort.

Fixes #79.