Handle missing title and/or action in user_log_events().
This happens when a log entry has its visibilty altered.  Note that
unlike hidden comments, which generate a 'hiddencomment' field, title
and/or action are just missing from the returned data with no explicit
"...hidden" field to replace it.  At least that's how it looks from
playing around with the API.  I can't find anywhere that actually
documents that.

Fixes #159.